+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 ## Master
 
+- Use swift instead of swiftc to fix compilation on Xcode 13 [@f-meloni][] - [#460](https://github.com/danger/swift/pull/460)
 - Add Homebrew path for Apple Silicon based macOS installations [@majd][] - [#458](https://github.com/danger/swift/pull/458)
 
 ## 3.10.2